Rename the dualpipe package to pipeline
