+++ AGENTS.md
@@ -73,7 +73,7 @@ python -m tools.memory_estimator --help   # peak-memory simulator for a given pa
 
 ## Architecture
 
-### DualPipeV Pipeline (`pithtrain/dualpipe/`)
+### DualPipeV Pipeline (`pithtrain/pipeline/`)
 
 The core pipeline assigns each rank two model chunks in a V-shape (the model is cut into `2 * pp_size` chunks; rank `r` holds chunks `r` and `2*pp_size-1-r`) and overlaps forward and backward execution across micro-batches. Each transformer layer is split into 5 stages:
 
@@ -83,7 +83,7 @@ The core pipeline assigns each rank two model chunks in a V-shape (the model is 
 4. **Combine** — All-to-all gather expert outputs (async on comm stream)
 5. **Aggregate** — Weighted expert output + residual connection
 
-Stages 1, 3, and 5 are the layer's compute entry points — `forward_stage1`, `forward_stage3`, `forward_stage5` on `LayerProtocol` (`pithtrain/models/interface.py`); stages 2 and 4 are all-to-all communication driven by the execution machinery (`pithtrain/dualpipe/execution.py`), not layer methods.
+Stages 1, 3, and 5 are the layer's compute entry points — `forward_stage1`, `forward_stage3`, `forward_stage5` on `LayerProtocol` (`pithtrain/models/interface.py`); stages 2 and 4 are all-to-all communication driven by the execution machinery (`pithtrain/pipeline/execution.py`), not layer methods.
 
 Key files:
 - `dualpipev.py` — Main scheduler and P2P between pipeline ranks: `DualPipeV.step()` orchestrates overlapped F/B across modules (supports `forward_only=True` for inference), `overlapped_forward_backward()` is the interleaved loop for one pair of micro-batches, plus `layer_partition()`, which distributes decoder layers across pipeline stages — edge stages (which hold `embed_tokens` / `norm`+`lm_head`) get fewer layers to balance memory.
