+++ NOTICE
@@ -34,7 +34,7 @@ SOFTWARE.
 The following files in this repository contain code derived from or closely
 adapted from the original DualPipe project:
 
-  pithtrain/dualpipe/dualpipev.py
+  pithtrain/pipeline/dualpipev.py
     - The DualPipeV class (scheduling algorithm, P2P communication methods,
       and state management) is derived from dualpipe/dualpipev.py in the
       original repository. The _append_irecv and _append_isend helpers are
@@ -42,13 +42,13 @@ adapted from the original DualPipe project:
       five-stage decomposition, the overlapped_forward_backward() loop, and
       the FSDP/FP8 integration are original additions.
 
-  pithtrain/dualpipe/execution.py
+  pithtrain/pipeline/execution.py
     - WeightGradStore and run_backward are derived from dualpipe/utils.py in
       the original repository. Everything else in the file -- the five-stage
       decomposition and its forward/backward implementations -- is original.
 
 A copy of the original MIT license is also retained at:
-  pithtrain/dualpipe/LICENSE
+  pithtrain/pipeline/LICENSE
 
 ================================================================================
