--- pithtrain/dualpipe/__init__.py
@@ -1,6 +1,0 @@
-from pithtrain.dualpipe.dualpipev import DualPipeV, Microbatch
-
-__all__ = [
-    "DualPipeV",
-    "Microbatch",
-]
