+++ pithtrain/modules/training.py
@@ -20,7 +20,6 @@ from transformers import AutoConfig
 
 from pithtrain.config import SlottedDefault
 from pithtrain.contexts import distributed, training
-from pithtrain.dualpipe import DualPipeV
 from pithtrain.models.deepseek_v2 import DeepSeekV2Model
 from pithtrain.models.gpt_oss import GptOssModel
 from pithtrain.models.qwen3_moe import Qwen3MoeModel
@@ -30,6 +29,7 @@ from pithtrain.modules.load_balance import force_balance, make_load_balance_loss
 from pithtrain.modules.optimizer import Muon
 from pithtrain.operators.grouped_linear import FP8GroupedLinear, GroupedLinear
 from pithtrain.operators.linear import FP8Linear
+from pithtrain.pipeline import DualPipeV
 
 from .distributed import DistributedCfg
