+++ pithtrain/operators/grouped_linear.py
@@ -6,7 +6,6 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from pithtrain.dualpipe.execution import WeightGradStore
 from pithtrain.operators.deepgemm_quantize import (
     fp8cast_blockwise_transpose_batched,
     fp8cast_rowwise_colwise,
@@ -14,6 +13,7 @@ from pithtrain.operators.deepgemm_quantize import (
 )
 from pithtrain.operators.fp8_weight_cache import FP8WeightCacheControl
 from pithtrain.operators.linear import ARCH_MAJOR
+from pithtrain.pipeline.execution import WeightGradStore
 
 
 class GroupedLinearFunc(torch.autograd.Function):
